+++ packages/camera/camera_web/CHANGELOG.md
@@ -1,8 +1,6 @@
 ## 0.3.5+4
 
-* Fixes a TypeError in `availableCameras` when a browser returns an invalid
-  `facingMode` capability.
-* Updates minimum supported SDK version to Flutter 3.35/Dart 3.9.
+* Updates minimum supported SDK version to Flutter 3.38/Dart 3.10.
 
 ## 0.3.5+3